requirements.txt
Django>=1.7
djangorestframework==3.2
